+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.0] - 2022-02-07
+
 * New: `wrap_text` for using custom control sequences.
 * Fix broken randomness in `html_id`, also using a fixed format (`[prefix]-[five digits]-[five digits]`)
 
@@ -47,7 +49,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Adds `paragraphize` filter
 * Adds `strip_control_characters` filter
 
-[Unreleased]: https://github.com/gglnx/twig-html-extended-extra/compare/v0.3.0...HEAD
+[Unreleased]: https://github.com/gglnx/twig-html-extended-extra/compare/v0.4.0...HEAD
+[0.4.0]: https://github.com/gglnx/twig-html-extended-extra/releases/tag/v0.4.0
 [0.3.0]: https://github.com/gglnx/twig-html-extended-extra/releases/tag/v0.3.0
 [0.2.4]: https://github.com/gglnx/twig-html-extended-extra/releases/tag/v0.2.4
 [0.2.3]: https://github.com/gglnx/twig-html-extended-extra/releases/tag/v0.2.3